Added ability to copy prisms; Added context menu to tree; boxpropertiesdialogbar hopefully fully functional;
git-svn-id: svn://136.177.114.72/svn_GW/WPhast/trunk@5543 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/ZoneResizeAction.h
@@ -24,8 +24,8 @@ protected:
 	CWPhastView   *m_pView;
 	CZoneActor    *m_pZone;
 	// vtkCubeSource *m_CubeSource;
-	float m_OriginalBounds[6];
-	float m_Bounds[6];
+	double m_OriginalBounds[6];
+	double m_Bounds[6];
 
 #ifdef USE_GET_LENGTHS_CENTER
 	// original values
